Tweaked track merging parameters.
With the previous commit, this show fantastic improvements on images
like photo6, river-3 and the dune series, at the cost of discarding
photo1 and photo9 (most likely because those are close to affine
projections).

In any case, improvements are so major that it's worth rejecting some
image series.

Even with tracks that have just 2 points, the results are so good that
most likely any errors are caused by incorrect fundamental matrices or
P3P using incorrect point matches.